+++ lib/script.sql
@@ -4,5 +4,6 @@ define DIST_UPLOAD = '&1'
 define DIST_FOLDER = '&2'
 define APP_ID = '&3'
 define API = '&4'
+define PLUGIN_NAME = '&5'
 
-script "&DIST_UPLOAD." "&DIST_FOLDER." &APP_ID. "&API."
+script "&DIST_UPLOAD." "&DIST_FOLDER." &APP_ID. "&API." "&PLUGIN_NAME."